requirements.txt
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
psycopg2==2.9.11
requests==2.33.1
